% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyrnr-package.R
\docType{package}
\name{tidyrnr-package}
\alias{tidyrnr}
\alias{tidyrnr-package}
\title{tidyrnr: Repeatability and reproducibility analysis}
\description{
Make R&R analysis easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lmaowisc/tidyrnr}
  \item \url{https://lmaowisc.github.io/tidyrnr/}
  \item Report bugs at \url{https://github.com/lmaowisc/tidyrnr/issues}
}

}
\author{
\strong{Maintainer}: Lu Mao \email{lmao@biostat.wisc.edu} (\href{https://orcid.org/0000-0002-8626-9822}{ORCID})

}
\keyword{internal}
